move codes to tablet_sink dir
