+++ be/src/exec/tablet_sink.cpp
@@ -50,8 +50,8 @@
 #include "config.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/stream_epoch_manager.h"
-#include "exec/tablet_sink_colocate_sender.h"
-#include "exec/tablet_sink_multi_sender.h"
+#include "exec/tablet_sink/tablet_sink_colocate_sender.h"
+#include "exec/tablet_sink/tablet_sink_multi_sender.h"
 #include "exprs/expr.h"
 #include "gutil/strings/fastmem.h"
 #include "gutil/strings/join.h"
