+++ be/src/exec/tablet_sink/tablet_sink_index_channel.cpp
@@ -32,7 +32,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "exec/tablet_sink_index_channel.h"
+#include "exec/tablet_sink/tablet_sink_index_channel.h"
 
 #include "agent/master_info.h"
 #include "agent/utils.h"
