+++ be/src/exec/tablet_sink/tablet_sink_multi_sender.h
@@ -48,8 +48,8 @@
 #include "common/tracer.h"
 #include "exec/data_sink.h"
 #include "exec/tablet_info.h"
-#include "exec/tablet_sink_index_channel.h"
-#include "exec/tablet_sink_sender.h"
+#include "exec/tablet_sink/tablet_sink_index_channel.h"
+#include "exec/tablet_sink/tablet_sink_sender.h"
 #include "gen_cpp/Types_types.h"
 #include "gen_cpp/doris_internal_service.pb.h"
 #include "gen_cpp/internal_service.pb.h"
